@@ -27,39 +32,56 @@
   "author": "Wix.com",
   "license": "MIT",
   "dependencies": {
-    "lodash": "^4.17.15",
-    "moment": "^2.24.0",
     "hoist-non-react-statics": "^3.3.1",
-    "react-native-swipe-gestures": "^1.0.5",
+    "immutable": "^4.0.0-rc.12",
+    "lodash": "^4.17.15",
+    "memoize-one": "^5.2.1",
     "prop-types": "^15.5.10",
-    "xdate": "^0.8.0",
-    "immutable": "^4.0.0-rc.12"
+    "react-native-swipe-gestures": "^1.0.5",
+    "recyclerlistview": "^3.0.5",
+    "xdate": "^0.8.0"
   },
   "devDependencies": {
     "@babel/core": "^7.12.9",
+    "@babel/eslint-parser": "^7.13.4",
     "@babel/runtime": "^7.12.5",
+    "@testing-library/react-native": "^11.0.0",
+    "@tsconfig/docusaurus": "^1.0.4",
+    "@types/lodash": "^4.14.170",
+    "@types/react": "^17.0.44",
+    "@types/react-native": "^0.64.19",
+    "@types/xdate": "^0.8.32",
+    "@typescript-eslint/eslint-plugin": "^2.13.0",
+    "@typescript-eslint/parser": "^2.13.0",
     "@welldone-software/why-did-you-render": "^6.0.3",
-    "babel-eslint": "^7.2.3",
+    "babel-eslint": "^10.1.0",
     "babel-jest": "^26.6.3",
-    "detox": "^16.7.2",
-    "eslint": "^6.8.0",
+    "detox": "^18.0.0",
+    "eslint": "^7.20.0",
     "eslint-config-prettier": "^6.11.0",
     "eslint-plugin-jest": "^24.1.3",
     "eslint-plugin-react": "^7.0.0",
+    "eslint-plugin-react-native": "^3.10.0",
     "husky": "^4.2.5",
     "jest": "^26.6.3",
     "jest-date-mock": "^1.0.8",
     "lint-staged": "^10.2.11",
-    "metro-react-native-babel-preset": "^0.64.0",
+    "metro-react-native-babel-preset": "^0.66.2",
     "mocha": "^7.1.0",
     "prettier": "^2.0.5",
-    "react": "16.11.0",
+    "react": "17.0.2",
     "react-component-driver": "^0.10.0",
-    "react-native": "0.62.2",
-    "react-native-navigation": "^6.7.1",
-    "react-test-renderer": "^17.0.1",
+    "react-native": "0.66.4",
+    "react-native-navigation": "^7.23.1",
+    "react-recipes": "^1.4.0",
+    "react-test-renderer": "^17.0.2",
+    "reassure": "^0.4.1",
     "semver": "5.x.x",
-    "shell-utils": "1.x.x"
+    "shell-utils": "1.x.x",
+    "typescript": "^4.3.3"
+  },
+  "optionalDependencies": {
+    "moment": "^2.24.0"
   },
   "jest": {
     "preset": "react-native",
