@@ -1,27 +1,27 @@
-const iface = require('./interface');
-const XDate = require('xdate');
+import {parseDate, xdateToData, toMarkingFormat} from './interface';
+import XDate from 'xdate';
 
-describe('calendar interface', () => {
-  describe('input', () => {
+describe('interface', () => {
+  describe('parseDate()', () => {
     it('should return undefined if date is undefined', () => {
-      const date = iface.parseDate();
+      const date = parseDate();
       expect(date).toBe(undefined);
     });
 
     it('should accept UTC timestamp as argument', () => {
-      const date = iface.parseDate(1479832134398);
+      const date = parseDate(1479832134398);
       expect(date.getTime()).toEqual(1479832134398);
       expect(date.getTimezoneOffset()).toEqual(0);
     });
 
-    it('should accept datestring as argument', () => {
-      const date = iface.parseDate('2012-03-16');
+    it('should accept dateString as argument', () => {
+      const date = parseDate('2012-03-16');
       expect(date.toString('yyyy-MM-dd')).toEqual('2012-03-16');
       expect(date.getTimezoneOffset()).toEqual(0);
     });
 
     it('should expect object with UTC timestamp as argument', () => {
-      const date = iface.parseDate({timestamp: 1479832134398});
+      const date = parseDate({timestamp: 1479832134398});
       expect(date.getTime()).toEqual(1479832134398);
       expect(date.getTimezoneOffset()).toEqual(0);
     });
@@ -35,7 +35,7 @@ describe('calendar interface', () => {
 
     it('should accept Date as argument', () => {
       const testDate = new Date(2015, 5, 5, 12, 0);
-      const date = iface.parseDate(testDate);
+      const date = parseDate(testDate);
       expect(date.toString('yyyy-MM-dd')).toEqual('2015-06-05');
     });
 
@@ -45,17 +45,17 @@ describe('calendar interface', () => {
         month: 5,
         day: 6
       };
-      const date = iface.parseDate(testDate);
+      const date = parseDate(testDate);
       expect(date.toString('yyyy-MM-dd')).toEqual('2015-05-06');
     });
   });
 
-  describe('output', () => {
+  describe('xdateToData()', () => {
     it('should convert xdate to data', () => {
       const time = 1479772800000;
       const testDate = XDate(time, true);
-      expect((testDate).toISOString()).toEqual('2016-11-22T00:00:00Z');
-      const data = iface.xdateToData(testDate);
+      expect(testDate.toISOString()).toEqual('2016-11-22T00:00:00Z');
+      const data = xdateToData(testDate);
       expect(data).toEqual({
         year: 2016,
         month: 11,
@@ -63,6 +63,14 @@ describe('calendar interface', () => {
         timestamp: 1479772800000,
         dateString: '2016-11-22'
       });
+    });
+  });
+
+  describe('toMarkingFormat()', () => {
+    it('should convert date to yyyy-MM-dd format string', () => {
+      const time = 1479772800000;
+      const testDate = XDate(time);
+      expect(toMarkingFormat(testDate)).toEqual(testDate.toString('yyyy-MM-dd'));
     });
   });
 });
