@@ -9,8 +9,8 @@ module.exports = {
     it: true,
     describe: true
   },
-  extends: ['eslint:recommended'],
-  parser: 'babel-eslint',
+  extends: ['eslint:recommended', 'plugin:@typescript-eslint/recommended'],
+  parser: '@typescript-eslint/parser',
   parserOptions: {
     ecmaFeatures: {
       experimentalObjectRestSpread: true,
@@ -18,13 +18,17 @@ module.exports = {
     },
     sourceType: 'module'
   },
-  plugins: ['react', 'jest'],
+  plugins: ['react', 'react-native', 'jest', '@typescript-eslint'],
   rules: {
+    'react-native/no-inline-styles': 1,
     'linebreak-style': ['error', 'unix'],
-    'no-unused-vars': 2,
+    'no-unused-vars': 1,
     'object-curly-spacing': ['error', 'never'],
     semi: ['error', 'always'],
     'react/jsx-uses-react': 2,
-    'react/jsx-uses-vars': 2
+    'react/jsx-uses-vars': 2,
+    '@typescript-eslint/no-use-before-define': 0,
+    '@typescript-eslint/no-var-requires': 0,
+    '@typescript-eslint/explicit-function-return-type': 0
   }
 };
