@@ -1,5 +1,6 @@
+import testIDs from '../example/src/testIDs';
+
 const {HEADER_MONTH_NAME} = require('../src/testIDs');
-const testIDs = require('../example/src/testIDs');
 
 describe('Horizontal Calendar List', () => {
   const FIRST_CALENDAR = `${testIDs.horizontalList.CONTAINER}_1526428800000`;
